# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Netflix API Documentation
  version: 1.0.0
  description: Comprehensive API documentation for all services in the Netflix API project.
servers:
  - url: http://localhost:4000
    description: Local server
paths:
...
  /exports/users:
    get:
      summary: Export all users
      description: Exports all user data.
      parameters:
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [json, xml]
      responses:
        '200':
          description: Export successful.
  /exports/profiles:
    get:
      summary: Export all profiles
      description: Exports all profile data.
      parameters:
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [json, xml]
      responses:
        '200':
          description: Export successful.
  /exports/subscriptions:
    get:
      summary: Export all subscriptions
      description: Exports all subscription data.
      parameters:
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [json, xml]
      responses:
        '200':
          description: Export successful.
...